refactor: move streamRoute into core so headless callers can use it
