+++ src/core/streamRoute.ts
